+++ libs/cart-store-credit/driver/magento/src/models/responses/remove-store-credit.type.ts
@@ -1,0 +1,7 @@
+import { MagentoCartWithStoreCredit } from './cart-with-store-credit.type';
+
+export interface MagentoRemoveStoreCreditResponse {
+  removeStoreCreditFromCart: {
+    cart: MagentoCartWithStoreCredit;
+  };
+}