+++ libs/cart-store-credit/state/src/reducers/store-feature-key.ts
@@ -1,0 +1,4 @@
+/**
+ * The key under which the cart store credit feature state will be stored.
+ */
+export const DAFF_CART_STORE_CREDIT_STORE_FEATURE_KEY = 'daffCartStoreCredit';