+++ libs/cart-store-credit/state/testing/src/state-testing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+
+import { DaffCartStoreCreditPageFacade } from '@daffodil/cart-store-credit/state';
+import { MockDaffOperationStateFacade } from '@daffodil/core/state/testing';
+
+/**
+ * Provides the {@link MockDaffCartStoreCreditPageFacade} for {@link DaffCartStoreCreditPageFacade}.
+ */
+@NgModule({
+  providers: [
+    { provide: DaffCartStoreCreditPageFacade, useExisting: MockDaffOperationStateFacade },
+  ],
+})
+export class DaffCartStoreCreditStateTestingModule {}