+++ libs/customer-order/driver/magento/src/driver.module.ts
@@ -4,14 +4,10 @@ import {
   ModuleWithProviders,
 } from '@angular/core';
 
-import { DAFF_MAGENTO_CACHEABLE_OPERATIONS } from '@daffodil/driver/magento';
 import { DaffOrderDriver } from '@daffodil/order/driver';
 
 import { DaffCustomerOrderMagentoService } from './order.service';
-import {
-  MAGENTO_GET_CUSTOMER_ORDERS_QUERY_NAME,
-  MAGENTO_GET_CUSTOMER_ORDER_QUERY_NAME,
-} from './queries/public_api';
+import { MagentoCustomerOrderCollectionTransformer } from './transforms/public_api';
 
 @NgModule({
   imports: [
@@ -23,6 +19,7 @@ export class DaffCustomerOrderMagentoDriverModule {
     return {
       ngModule: DaffCustomerOrderMagentoDriverModule,
       providers: [
+        MagentoCustomerOrderCollectionTransformer,
         {
           provide: DaffOrderDriver,
           useExisting: DaffCustomerOrderMagentoService,