+++ libs/customer-store-credit/driver/magento/src/models/responses/customer-order-with-store-credit.type.ts
@@ -1,0 +1,8 @@
+import { MagentoCustomerOrder } from '@daffodil/customer-order/driver/magento';
+import { MagentoMoney } from '@daffodil/driver/magento';
+
+export interface MagentoCustomerOrderWithStoreCredit extends MagentoCustomerOrder {
+  total: MagentoCustomerOrder['total'] & {
+    total_store_credit: MagentoMoney;
+  };
+}