+++ libs/customer-store-credit/driver/testing/src/drivers/public_api.ts
@@ -1,0 +1,2 @@
+export { DaffCustomerStoreCreditTestingDriverModule } from './driver.module';
+export { DaffCustomerStoreCreditTestingDriver } from './store-credit.service';