+++ libs/customer-store-credit/routing/src/guards/public_api.ts
@@ -1,0 +1,1 @@
+export { DaffCustomerStoreCreditClearErrorsGuard } from './clear-errors.guard';